+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Declare official support of Python 3.12
 * Cache-Control option when uploading files
+* Add `retry-for` parameter to retry for some period (in minutes) when an upload or download fails, instead of hardcoded 5 retries 
 
 ### Infrastructure
 * Remove unsupported PyPy 3.7 from tests matrix and add PyPy 3.10 instead